feat(core): rework property map-like interface with boost flat_map
